pull out google vertex functionality to client file
