+++ packages/lesswrong/server.ts
@@ -80,7 +80,7 @@ import "./server/scripts/generativeModels/testModGPTOnComments";
 import "./server/scripts/generativeModels/coverImage";
 import "./server/scripts/addManualReviewArt";
 import "./server/scripts/backfillRecombee";
-import "./server/scripts/backfillBigquery";
+import "./server/scripts/backfillGoogleVertex";
 // doesn't pass unit tests but works fine. Leaving commented out for now
 // import './server/scripts/generativeModels/autoSpotlight';
 import "./server/scripts/mongoQueryToSQL";
