+++ packages/lesswrong/server/databaseSettings.ts
@@ -95,3 +95,7 @@ function validateSettings(registeredSettings: Record<string, "server" | "public"
 
 export const openAIApiKey = new DatabaseServerSetting<string|null>('languageModels.openai.apiKey', null);
 export const openAIOrganizationId = new DatabaseServerSetting<string|null>('languageModels.openai.organizationId', null);
+
+export const vertexDocumentServiceParentPathSetting = new DatabaseServerSetting<string | null>('googleVertex.documentServiceParentPath', null);
+export const vertexUserEventServiceParentPathSetting = new DatabaseServerSetting<string | null>('googleVertex.userEventServiceParentPath', null);
+export const vertexRecommendationServingConfigPathSetting = new DatabaseServerSetting<string | null>('googleVertex.recommendationServingConfigPath', null);
